added hibernate validation
