+++ src/main/webapp/WEB-INF/classes/resources/domains.xml
@@ -1,20 +1,20 @@
 <?xml version='1.0' encoding='utf-8'?>
-<communities>
-    <community>
+<domains>
+    <domain>
         <name>Atmospheric Sciences</name>
         <platform>Single Station</platform>
         <platform>Radiosonde</platform>
         <platform>Wind Profiler</platform>
         <platform>Aircraft</platform>
         <platform>Glider</platform>
-    </community>
-    <community>
+    </domain>
+    <domain>
         <name>Physical Ocean Sciences</name>
         <platform>Moored Buoy</platform>
         <platform>CTD Sonde</platform>
-    </community>
-    <community>
+    </domain>
+    <domain>
         <name>Bio-Logging</name>
         <platform>eTag</platform>
-    </community>
-</communities>
+    </domain>
+</domains>
