--- src/main/webapp/WEB-INF/classes/resources/sql/embeddedDerby/db.properties
@@ -1,4 +1,0 @@
-driver=org.apache.derby.jdbc.EmbeddedDriver
-databaseName=rosettaDb
-username=
-password=
